requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Authentication & Security
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0
boto3==1.34.0

# HTTP & Validation
python-multipart==0.0.6

# Production
gunicorn==21.2.0
